Added much more stablitiy to the application Stock history prices load correctly now and can deal with dups etc Schema has been updated with more unique constraints and tighter data types Added a testing util for control of data

+++ src/main/java/com/nevex/roboinvesting/database/StockPricesRepository.java
@@ -1,8 +1,11 @@
 package com.nevex.roboinvesting.database;
 
 import com.nevex.roboinvesting.database.entity.StockPricesEntity;
+import com.nevex.roboinvesting.database.entity.StockPricesHistoricalEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigInteger;
 import java.util.Optional;
@@ -11,7 +14,7 @@ import java.util.Optional;
  * Created by Mark Cunningham on 8/8/2017.
  */
 @Repository
-public interface StockPricesRepository extends CrudRepository<StockPricesEntity, BigInteger> {
+public interface StockPricesRepository extends CrudRepository<StockPricesEntity, Integer> {
 
     Optional<StockPricesEntity> findBySymbol(String symbol);
